hotfix : float값 변환 오류 수정

+++ src/utils/floatHexValue.js
@@ -1,7 +1,7 @@
 export default function floatHexValue(arrCsvContentHex, value) {
   const valueHex = parseFloat(value);
 
-  let floatArr = new Float32Array(1);
+  let floatArr = new Float64Array(1);
   floatArr[0] = valueHex;
 
   let barr = new Int8Array(floatArr.buffer);
